move time functions out to a TimeframeUtils file
