+++ client/js/app/utils/FormatUtils.js
