+++ client/js/app/validations/ExplorerValidations.js
@@ -1,7 +1,7 @@
 var _ = require('lodash');
 var filterValidations = require('../validations/FilterValidations').filter;
 var runValidations = require('../utils/ValidationUtils').runValidations;
-var ExplorerUtils = require('../utils/ExplorerUtils');
+var TimeframeUtils = require('../utils/TimeframeUtils');
 var FilterUtils = require('../utils/FilterUtils');
 
 module.exports = {
@@ -57,13 +57,13 @@ module.exports = {
     time: {
       validator: function(explorer) {
         var time = explorer.query.time || {};
-        if (ExplorerUtils.timeframeType(explorer.query.time) === 'relative') {
+        if (TimeframeUtils.timeframeType(explorer.query.time) === 'relative') {
           if (time.relativity && time.amount && time.sub_timeframe) {
             return true;
           } else {
             return "You must choose all 3 options for relative timeframes.";
           }
-        } else if (ExplorerUtils.timeframeType(explorer.query.time) === 'absolute') {
+        } else if (TimeframeUtils.timeframeType(explorer.query.time) === 'absolute') {
           if (time.start && time.end) {
             return true;
           } else {
